+++ packages/vsce/src/doc/resources/IJVMServer.ts
@@ -1,0 +1,6 @@
+import { IResource } from "./IResource";
+
+export interface IJVMServer extends IResource {
+    name: string;
+    enableStatus: string;
+}