+++ packages/vsce-api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the IBM CICS for Zowe Explorer API will be documented in this file.
 
+## Recent Changes
+
+- Added new resource type - JVM server
+
 ## `6.10.0`
 
 - Populate supported resources list